third_party/WebKit/LayoutTests/external/wpt/css/css-backgrounds/background-332-expected.txt

This is a testharness.js-based test.
FAIL background_specified_image assert_equals: background specified value for background-image expected "url(support/60x60-green.png)" but got "url(\"http://web-platform.test:8001/css/css-backgrounds/support/60x60-green.png\")"
PASS background_specified_position
FAIL background_specified_size assert_equals: background specified value for background-size expected "10em 10em" but got "160px"
FAIL background_specified_repeat assert_equals: background specified value for background-repeat expected "round round" but got "round"
PASS background_specified_attachment
PASS background_specified_origin
PASS background_specified_clip
PASS background_specified_color
Harness: the test ran to completion.
